fix(models): publish-review backlog — legacy root layout, spawn wedge, probe cache, collapse fixes
Remaining review backlog, backend side:

- POST /models/upload gets its frozen on-Hub shape back: root_layout=True
  pushes the final checkpoint's files to the repo ROOT (loadable by a plain
  from_pretrained), no pin, no card index, and the return is exactly the
  pre-multi-checkpoint {repo_id, url, tags}. The step-addressed layout had
  silently replaced it, leaving legacy pushes without a root config.json.
- ModelUploadManager.start no longer wedges at "running" when the worker
  thread fails to spawn — the slot lands on a visible "error" and the route
  maps it to a 500 (the wedge would also have jammed the new delete guard).
- _published_repo_state gains a 30s TTL cache so the picker's per-dialog-open
  Hub probe stops costing a roundtrip each time; failures are never cached,
  and the post-upload reader bypasses with fresh=True.
- The run→hub listing collapse matches repo ids case-insensitively (the same
  fold as every jobs-side dedup) and now carries dataset_episodes, which the
  collapse tuple silently dropped from a published run's row.
- set_hf_repo_id follows the zero-then-restamp queue_position protocol that
  rename/start/reorder_queue established, so a stamped read can't freeze into
  job.json.
- tests: test_pr61_adversarial.py folds into test_models.py (nobody hunting
  models coverage would find a "pr61" file), its dead patch-if-False construct
  is gone, and the typos hook's `anothers` failure is fixed. New coverage for
  every fix above; the mis-typed test id is renamed.
